net.unbound: Merge luaunbound and prosody defaults in absence of user config (fixes #1763) (thanks rgd)
add_defaults() is supposed to merge 3 tables, the defaults in
luaunbound, the defaults from prosody and any config from the prosody
config file.  In the case where no `unbound={}` has been in the config,
it skips over the merge and returns only the prosody built-in defaults.

This results in libunbound skipping reading resolv.conf and uses its
default behavior of full recursive resolution.

Prior to #1737 there were only two tables, the luaunbound defaults and
the prosody config, where bypassing the merge and returning the former
did the right thing.