mod_mam: Store archives with sub-second precision timestamps
Changes sub-second part of example timestamp to .5 in order to avoid
floating point issues.

Some clients use timestamps when ordering messages which can lead to
messages having the same timestamp ending up in the wrong order.

It would be better to preserve the order messages are sent in, which is
the order they were stored in.