Backed out changeset 1bc2220cd6ec
The use of the error helpers creates an `<error/>` child element
containing the error condition. This is however not allowed as per
XEP-0198, which specifies that the error condition is to be a direct
child of the `<failed/>` stream management element.

This has triggered a fun reconnect loop in aioxmpp where it was
reported by a user [1].

   [1]: https://github.com/horazont/aioxmpp/issues/382