mod_smacks: Change boolean attribute from '1' to 'true' for compatibility
Conversations 2.10.10 and earlier expect this to be literally 'true' and don't
recognise '1'. This leads to it not attempting resumption with Prosody at all
since this change was introduced in 36ba170c4fd0.

Thanks to Zash for noticing, debugging and diagnosing this issue.

This issue is fixed in Conversations commit 052c58f3 (unreleased at the time
of writing).