util.sasl.oauthbearer: Return username from callback instead using authzid (BC)
RFC 6120 states that
>  If the initiating entity does not wish to act on behalf of another
>  entity, it MUST NOT provide an authorization identity.

Thus it seems weird to require it here.  We can instead expect an
username from the token data passed back from the profile.

This follows the practice of util.sasl.external where the profile
callback returns the selected username, making the authentication module
responsible for extracting the username from the token.