moduleapi: Add 'peek' to :may() and new :could() helper to suppress logging
The current method logs scary "access denied" messages on failure - this is
generally very useful when debugging access control stuff, but in some cases
the call is simply a check to see if someone *could* perform an action, even
if they haven't requested it yet. One example is determining whether to show
the user as an admin in disco.

The 'peek' parameter, if true, will suppress such logging.

The :could() method is just a simple helper that can make the calling code a
bit more readable (suggested by Zash).