util.startup: Add prosody.started promise to easily execute code after startup
To avoid a race where server-started fires before the promise function body is
run (on next tick), I moved server-started to fire on the next tick, which
seems sensible anyway.

Errors are logged, I'm not sure if we ought to be doing something more here.
I'm sure we'll find out.