util.dependencies: Print tables itself to reduce number of imports
Rationale: See diffstat

When this module is imported, it ends up calling stty via term_width()
in util.human.io.table(). When this happens outside of a terminal, the
following message is sent to stdout:
	stty: 'standard input': Inappropriate ioctl for device

Not importing this module avoids that.

Furthermore three is value in this module having minimal dependencies as
they might not be available when it does the checks.

Ref a1fed82c44b9