mod_http: Use RFC 7239 Forwarded header to find original client IP
Prefer over X-Forwarded-* since it has an actual specification.

Main practical difference is that Forwarded may carry more properties
than only the IP address since it is a structured header.

Since we parse it into an array, it is easier to do the logical thing
and iterate backwards trough proxies until an untrusted one is
encountered. Compare the handling of X-Forwarded-For.

The 'secure' field now accounts for the full chain of proxies, which
must be secure all the way to be considered secure.