mod_http: Fix error if 'access_control_allow_origins' is set
Because it changes the type of the 'opt_origins' variable from util.set
to the internal _items table so next time an http app is added an error
"attempt to call a nil value (method 'empty')" is triggered.  The value
is not used anywhere else.

Noticed when reviewing uses of the '_items' set property.

Not reported by any users, implying this setting is rarely used.