mod_muc_mam: Improve wording of enable setting
Suggested by jstein in the chat

This option label is used by XMPP clients to explain what the option does.
a) The user should know where the data is archived.
b) The user needs a statement that can be enabled/disabled by the variable. A question would have the wrong logic here.