util.prosodyctl: Use notify socket to wait for Prosody to be ready
Previously, prosodyctl only waits for the pidfile to appear, which does
not necessarily mean that Prosody is fully ready to receive traffic.

By waiting until Prosody says it's ready via the systemd notify socket
we know for sure that Prosody is really ready.

Notably this should ensure that when running `make integration-test`
Prosody is really ready when Scansion starts running tests.

Not sure if this timeout handling is optimal.