mod_s2s: Limit size of outgoing stanza queue
This queue is used to buffer stanzas while waiting for an outgoing s2s
connection to be established.

Limit it to prevent excessive memory usage.

Default chosen to approximate how many average stanzas fits in the
server_epoll default max_send_buffer_size of 32 MiB

Returns a custom error instead of the default core.stanza_router
"Communication with remote domains is not enabled" from is sent back,
which does not describe what is happening here.

Closes #1106