mod_authz_internal: Make 'prosody:guest' default role for all unknown JIDs
This fixes an issue where e.g. remote users or even other users on the server
were unable to list MUC rooms.

We want to define a permission to list MUC rooms, but we want it to be
available to everyone by default (the traditional behaviour).

prosody:guest is the lowest role we have. I ran a quick check and it isn't
really used for anything right now that would be concerning.

It was originally designed for anonymous logins. I think it's safe to treat
remote JIDs as equivalent, since we have no trust relationship with anonymous
users either.