core.moduleapi: Include source modules when handling items
This improves consistency. Previously the 'source' field was only
provided in the original event when an item was added. It is used to
report the name of the module providing the item in a few places.

Also considered adding a new API to modulemanager returning a mapping
of items per module and then using that here.