mod_storage_sql: Detect SQLite3 without UPSERT (or SQLCipher 3.x)
SQLCipher v3.4.1 (the version in Debian 12) is based on SQLite3 v3.15.2,
while UPSERT support was introduced in SQLite3 v3.24.0

This check was not needed before because we v3.24.0 has not been in a
version of Debian we support for a long, long time.

Note however that SQLCipher databases are not compatible across major
versions, upgrading from v3.x to v4.x requires executing a migration.
Attempts at making `prosodyctl mod_storage_sql upgrade` perform such a
migration has not been successful.

Executing the following in the `sqlcipher` tool should do the migration:

	PRAGMA key = '<key material>';
	PRAGMA cipher_migrate;