prosodyctl: Further deprecate start/stop/restart commands when installed
Despite the warning we introduced, many people continue to try using
prosodyctl to manage Prosody in the presence of systemctl (e.g. #1688).

Also, despite the warning, prosodyctl proceeded with the operation. This means
the commands could be invoked by accident, and cause a situation that is hard
to recover from (needing to manually track down stray processes).

This commit disables all the problematic commands by default, but this can
still be overridden using --force or via a config option.

We only perform this check when we believe Prosody has been "installed" for
system-wide use (i.e. running it from a source directory is still supported).