util.sasl: Preserve 'userdata' field between clones
The :clean_clone() method is designed to provide a new cloned SASL handler,
to be used when starting a fresh SASL negotiation on an existing connection.

The userdata field is currently populated by mod_saslauth with the "read-only"
information that the channel binding methods need to do their stuff.

When :clean_clone() does not preserve this, it causes tracebacks in the cb
profile handlers due to the property being nil.

This does mean that SASL handlers should now not be reused (even when cloned)
across different connections, if they ever could.